chore: Moved exporters logic and their references to be conatined within the graph module
