+++ ra2ce/graph/networks.py
@@ -34,11 +34,11 @@ import pyproj
 from shapely.geometry import MultiLineString
 from ra2ce.graph.network_config_data.network_config_data import NetworkConfigData
 
-import ra2ce.graph.networks_utils as nut
+from ra2ce.graph import networks_utils as nut
 from ra2ce.graph.segmentation import Segmentation
 from ra2ce.io.readers import GraphPickleReader
 from ra2ce.io.writers import JsonExporter
-from ra2ce.io.writers.network_exporter_factory import NetworkExporterFactory
+from ra2ce.graph.exporters.network_exporter_factory import NetworkExporterFactory
 
 
 class Network:
