+++ tests/graph/exporters/test_geodataframe_network_exporter.py
@@ -3,8 +3,8 @@ import shutil
 import pytest
 from geopandas import GeoDataFrame
 
-from ra2ce.io.writers.geodataframe_network_exporter import GeoDataFrameNetworkExporter
-from ra2ce.io.writers.network_exporter_base import NetworkExporterBase
+from ra2ce.graph.exporters.geodataframe_network_exporter import GeoDataFrameNetworkExporter
+from ra2ce.graph.exporters.network_exporter_base import NetworkExporterBase
 from tests import test_results
